FIX : Samples : Updated after package refactoring
