+++ src/apps/kotlin/slatekit-sampleapps/sampleapp-core/src/main/kotlin/slatekit/sampleapp/core/apis/SampleErrorsApi.kt
@@ -2,6 +2,7 @@ package slatekit.sampleapp.core.apis
 
 import slatekit.apis.support.ApiWithMiddleware
 import slatekit.common.*
+import slatekit.common.requests.Request
 import slatekit.common.results.ResultCode.UNEXPECTED_ERROR
 import slatekit.common.results.ResultFuncs.badRequest
 import slatekit.common.results.ResultFuncs.failure
