+++ src/apps/kotlin/slatekit-sampleapps/sampleapp-core/src/main/kotlin/slatekit/sampleapp/core/apis/SamplePOKOApi.kt
@@ -1,9 +1,9 @@
 package slatekit.sampleapp.core.apis
 
 import slatekit.common.DateTime
-import slatekit.common.Request
 import slatekit.common.Result
 import slatekit.common.ResultMsg
+import slatekit.common.requests.Request
 import slatekit.common.results.ResultFuncs.badRequest
 import slatekit.common.results.ResultFuncs.failure
 import slatekit.common.results.ResultFuncs.success
