+++ src/apps/kotlin/slatekit-sampleapps/sampleapp-core/src/main/kotlin/slatekit/sampleapp/core/common/AppEncryptor.kt
@@ -11,7 +11,8 @@
 
 package slatekit.sampleapp.core.common
 
+import slatekit.common.encrypt.B64Java8
 import slatekit.common.encrypt.Encryptor
 
-object AppEncryptor : Encryptor("UrbanDonut201607", "25SK1CA6F75827B5"){
+object AppEncryptor : Encryptor("UrbanDonut201607", "25SK1CA6F75827B5", B64Java8){
 }
